dom/base/test/test_bug1314032.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1314032</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1314032">Mozilla Bug 1243846</a>
<p id="display"></p>
<pre id="test">
<script type="application/javascript">

	let win = window.open(URL.createObjectURL(new Blob([
		'<meta charset="utf-8">' +
		'<script>' +
		  'let observer = new IntersectionObserver(([entry]) => {' +
		    'document.body.textContent += entry.time' +
		  '});' +
			'observer.observe(document.documentElement);' +
		'<\/script>'
	], {'type': 'text/html'})));
	
	win.onload = function () {
		win.close();
		ok(true);
		SimpleTest.finish();
	}
	
	SimpleTest.waitForExplicitFinish();

</script>
</pre>
<div id="log">
</div>
</body>
</html>
